pyproject.toml
[project]
name = "savdhaan-ai"
version = "0.1.0"
description = "AI risk copilot — scam detection and beyond"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "Savdhaan AI Team" }]
dependencies = [
    # Web framework
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "gunicorn>=23.0,<24.0",
    # Database
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.30,<1.0",
    "alembic>=1.14,<2.0",
    # Redis
    "redis>=5.0,<6.0",
    # Task queue
    "celery[redis]>=5.4,<6.0",
    # HTTP client
    "httpx>=0.28,<1.0",
    # Validation & settings
    "pydantic>=2.10,<3.0",
    "pydantic-settings>=2.7,<3.0",
    # AI
    "anthropic>=0.42,<1.0",
    # Auth & security
    "bcrypt>=4.2,<5.0",
    "pyjwt[crypto]>=2.10,<3.0",
    # Utilities
    "python-multipart>=0.0.18",
    "pillow>=11.0,<12.0",
    "phonenumbers>=8.13,<9.0",
    "python-whois>=0.9,<1.0",
    "tldextract>=5.1,<6.0",
    # OCR fallback
    "pytesseract>=0.3,<1.0",
    # Logging
    "structlog>=24.4,<25.0",
    # Prometheus metrics
    "prometheus-client>=0.21,<1.0",
    "starlette-prometheus>=0.10,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-cov>=6.0,<7.0",
    "pytest-httpx>=0.34,<1.0",
    "httpx>=0.28,<1.0",
    "ruff>=0.9,<1.0",
    "mypy>=1.14,<2.0",
    "pre-commit>=4.0,<5.0",
    "factory-boy>=3.3,<4.0",
    "faker>=33.0,<34.0",
    "aiosqlite>=0.20,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "T20",  # flake8-print
    "S",    # flake8-bandit (security)
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "RET",  # flake8-return
    "ARG",  # flake8-unused-arguments
]
ignore = [
    "S101",  # assert allowed in tests
    "B008",  # function call in default arg (FastAPI Depends)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG001", "ARG002"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
